+++ plt/plt-block-state/src/entity/block_state/p9.rs
@@ -105,30 +105,4 @@ impl BlockStateP9 {
     }
 }
 
-impl Accounts for BlockStateP9 {
-    fn account_by_address<C: EntityContextTypes>(
-        &self,
-        context: &EntityContext<C>,
-        address: &AccountAddress,
-    ) -> Result<Account, AccountNotFoundByAddressError> {
-        let account_index = context.external.account_index_by_account_address(address)?;
-        Ok(Account::from_existing_account(account_index))
-    }
 
-    fn account_by_index<C: EntityContextTypes>(
-        &self,
-        context: &EntityContext<C>,
-        account_index: AccountIndex,
-    ) -> Result<AccountWithCanonicalAddress, AccountNotFoundByIndexError> {
-        let canonical_account_address = context
-            .external
-            .account_canonical_address_by_account_index(account_index)?;
-
-        let account = Account::from_existing_account(account_index);
-
-        Ok(AccountWithCanonicalAddress {
-            account,
-            canonical_account_address,
-        })
-    }
-}
