+++ plt/plt-scheduler/src/block_state_traits.rs
@@ -1,1 +1,3 @@
-mod tokens;
+//! Traits implemented for block state types with the purpose of making the scheduler code generic.
+
+pub mod accounts;
