--- plt/plt-scheduler/src/block_state_traits/tokens.rs
@@ -1,17 +1,0 @@
-/// Trait that defines block state operations related to support for P9 tokens.
-pub trait TokensP9 {
-    /// Lookup the account using an account address.
-    fn account_by_address<C: EntityContextTypes>(
-        &self,
-        context: &EntityContext<C>,
-        address: &AccountAddress,
-    ) -> Result<Account, AccountNotFoundByAddressError>;
-
-    /// Lookup the account using an account index. Returns both the opaque account
-    /// representation and the account canonical address.
-    fn account_by_index<C: EntityContextTypes>(
-        &self,
-        context: &EntityContext<C>,
-        account_index: AccountIndex,
-    ) -> Result<AccountWithCanonicalAddress, AccountNotFoundByIndexError>;
-}
